pyproject.toml
[tool.poetry]
name = "datafram_test"
version = "0.1.0"
description = ""
authors = ["Lex Chen <clang2you@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.5"
pyqt5 = "^5.15.1"
pyqt5-tools = "^5.15.1"
pyinstaller = "^4.1"
configparser = "^5.0.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
